Removed glem_param_groups from cmd line; moved GLEMOptimizer
