+++ python/graphstorm/model/gnn.py
@@ -133,40 +133,6 @@ class GSOptimizer():
                     if isinstance(v, th.Tensor):
                         state[k] = v.to(device)
 
-class GLEMOptimizer(GSOptimizer):
-    """ An optimizer specific for GLEM, implementing on/off switches of sparse optimizer.
-    """
-    def _clear_traces(self):
-        """ Clear the traces in sparse optimizers.
-        """
-        for optimizer in self.sparse_opts:
-            for emb in optimizer._params:
-                emb.reset_trace()
-
-    def zero_grad(self, optimize_sparse_params=True):
-        """ Setting the gradient to zero
-        """
-        all_opts = self.dense_opts + self.lm_opts
-        if optimize_sparse_params:
-            all_opts += self.sparse_opts
-        for optimizer in all_opts:
-            optimizer.zero_grad()
-        if not optimize_sparse_params:
-            # force reset trace for sparse opt when sparse emb are frozen
-            # under this condition, emb still collects traces with grad being None's.
-            # we need to do this to ensure the gradient update step after unfreezing
-            # can be performed correctly.
-            self._clear_traces()
-
-    def step(self, optimize_sparse_params=True):
-        """ Moving the optimizer
-        """
-        all_opts = self.dense_opts + self.lm_opts
-        if optimize_sparse_params:
-            all_opts += self.sparse_opts
-        for optimizer in all_opts:
-            optimizer.step()
-
 class GSgnnModelBase(nn.Module):
     """ GraphStorm GNN model base class
